Fix `-static-pie` bug, add Constructors/ and Firmware/, etc.
(Thank you to monkuous on the OSDev Discord for helping me with this <3, this is so obscure I don't think I would've figured it out otherwise)

It turns out -static-pie isn't really what I need, what I'm looking for is -pie -Wl,--no-dynamic-linker.

(From my understanding, `-static-pie` is often broken, especially on cross compilers, but it does so silently, so you don't even know anything is wrong - this only properly shows up with global variables though, and only on ELF (PE was fine))

(Since `-static-pie` doesn't produce a _DYNAMIC section, it has no reference for R_X86_64_RELATIVE, and assumes it to be zero)

I also added a few other features (and an additional check); namely, the kernel now has:

-> Constructors/, which for now only sets up EFI variables, but in the future should set up drivers and such.

-> Firmware/, which contains firmware-specific definitions and such.

(I also made EntrypointStatusMessage static, so it doesn't cause issues)

+++ Boot/Efi/Efi/Efi.h
@@ -32,7 +32,7 @@
 
   typedef uint64 efiPhysicalAddress;
   typedef uint64 efiVirtualAddress;
-  typedef uint64 efiNotImplemented; // sizeof(uint64*) is still sizeof(void*), so this is okay
+  typedef uintptr efiNotImplemented; // sizeof(uintptr) == sizeof(void*), so this is okay
 
   typedef struct _efiInputKey {
